move BuildAddLabelsAndSchemaToImage to image pkg
this was the only caller, and it fits better alongside the other main Build function
