+++ pkg/image/build.go
@@ -16,7 +16,6 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 
 	"github.com/replicate/cog/pkg/config"
-	"github.com/replicate/cog/pkg/docker"
 	"github.com/replicate/cog/pkg/docker/command"
 	"github.com/replicate/cog/pkg/dockercontext"
 	"github.com/replicate/cog/pkg/dockerfile"
@@ -272,8 +271,27 @@ func Build(ctx context.Context, cfg *config.Config, dir, imageName string, secre
 		labels[key] = val
 	}
 
-	if err := docker.BuildAddLabelsAndSchemaToImage(ctx, dockerCommand, imageName, labels, bundledSchemaFile); err != nil {
+	if err := BuildAddLabelsAndSchemaToImage(ctx, dockerCommand, imageName, labels, bundledSchemaFile); err != nil {
 		return fmt.Errorf("Failed to add labels to image: %w", err)
+	}
+	return nil
+}
+
+// BuildAddLabelsAndSchemaToImage builds a cog model with labels and schema.
+//
+// The new image is based on the provided image with the labels and schema file appended to it.
+func BuildAddLabelsAndSchemaToImage(ctx context.Context, dockerClient command.Command, image string, labels map[string]string, bundledSchemaFile string) error {
+	dockerfile := "FROM " + image + "\n"
+	dockerfile += "COPY " + bundledSchemaFile + " .cog\n"
+
+	buildOpts := command.ImageBuildOptions{
+		DockerfileContents: dockerfile,
+		ImageName:          image,
+		Labels:             labels,
+	}
+
+	if err := dockerClient.ImageBuild(ctx, buildOpts); err != nil {
+		return fmt.Errorf("Failed to add labels and schema to image: %w", err)
 	}
 	return nil
 }
